Began integrating changes from recent end-to-end spike
Use the messagePicker in SlotBodyEnd
